func: add checking discard pile for cards
Added method to allow players to grab from the discard pile.
Works by looking at all other players and seeing if the discarded
card is beneficial to them: if it completes one of their possibles.
Also, if the player who picked up from the discard is next, skips
grabbing from the deck for their next turn.